requirements.txt
pandas
pyHIve
matplotlib
textblob
preprocessor
streamlit
plotly
wordcloud
nltk
Image
pillow

